Added items to collect and avoid
Added the ball and bomb to the game as things the player need to collect and avoid when player. I also add more time to the timer and made a game over scene. If the user lets go of the screen or hits a bomb they go to that screen.